// apps/api/generated/prisma/enums.ts
export const AudienceType = {
  INVESTORS: 'INVESTORS',
  CUSTOMERS: 'CUSTOMERS',
  EXECUTIVES: 'EXECUTIVES',
  TEAM: 'TEAM',
  CONFERENCE: 'CONFERENCE',
  BOARD: 'BOARD',
  TECHNICAL: 'TECHNICAL'
} as const

export type AudienceType = (typeof AudienceType)[keyof typeof AudienceType]


export const PitchGoal = {
  RAISE_FUNDING: 'RAISE_FUNDING',
  SELL_PRODUCT: 'SELL_PRODUCT',
  GET_BUYIN: 'GET_BUYIN',
  EDUCATE: 'EDUCATE',
  INSPIRE: 'INSPIRE',
  REPORT_RESULTS: 'REPORT_RESULTS'
} as const

export type PitchGoal = (typeof PitchGoal)[keyof typeof PitchGoal]


export const ToneStyle = {
  FORMAL: 'FORMAL',
  CONVERSATIONAL: 'CONVERSATIONAL',
  BOLD: 'BOLD',
  INSPIRATIONAL: 'INSPIRATIONAL',
  ANALYTICAL: 'ANALYTICAL',
  STORYTELLING: 'STORYTELLING'
} as const

export type ToneStyle = (typeof ToneStyle)[keyof typeof ToneStyle]


export const CompanyStage = {
  IDEA: 'IDEA',
  MVP: 'MVP',
  GROWTH: 'GROWTH',
  ENTERPRISE: 'ENTERPRISE'
} as const

export type CompanyStage = (typeof CompanyStage)[keyof typeof CompanyStage]


export const TechnicalLevel = {
  NON_TECHNICAL: 'NON_TECHNICAL',
  SEMI_TECHNICAL: 'SEMI_TECHNICAL',
  TECHNICAL: 'TECHNICAL',
  HIGHLY_TECHNICAL: 'HIGHLY_TECHNICAL'
} as const

export type TechnicalLevel = (typeof TechnicalLevel)[keyof typeof TechnicalLevel]